Correction of remaining typing errors.

See merge request proj104/2024-2025/python-poly-opt!8

+++ src/ncpol3sdpa/rules.py
@@ -10,7 +10,7 @@ class Rule:
         """Private method creating a Tuple of equivalence"""
 
         # express polynomial as a list of monomial
-        polynomial: List[Tuple[Tuple[int, ...], int]] = poly(
+        polynomial: List[Tuple[Tuple[int, ...], float]] = poly(
             constraint.polynomial
         ).terms()
 
@@ -50,7 +50,7 @@ def apply_rule(monomial: Expr, rules: Dict[Expr, Expr]) -> Expr:
 def apply_rule_to_polynomial(polynomial: Expr, rules: Dict[Expr, Expr]) -> Expr:
     """Apply a rule to a polynomial"""
 
-    poly_dict: Dict[Expr, int] = polynomial.as_coefficients_dict()  # type: ignore
+    poly_dict: Dict[Expr, float] = polynomial.as_coefficients_dict()
 
     res: Expr = S.Zero
     for monomial, coef in poly_dict.items():
